Renamed the audit publisher to AuditEventPublisher
